requirements.txt
flask
flask-cors
flask-jwt-extended
pymongo[srv]
bcrypt
python-dotenv
requests
gunicorn
